requirements.txt
streamlit==1.41.1
plotly==5.15.0
ccxt==4.0.91
numpy==1.26.0
pandas==2.1.1
requests==2.31.0
python-dotenv